refactor(pips): move backend helpers to linopy/backends/pips.py
Relocate the top-level linopy/pips.py into a new linopy/backends/ subpackage
so backend-specific code no longer sits at the package root (the PIPS solver
class stays in solvers.py alongside the others). Public API is now
linopy.backends.pips.*; update the two internal imports, tests and docs.
