+++ doc/release_notes.rst
@@ -21,7 +21,7 @@ Upcoming Version
 
 * New ``Model.to_pips_files`` exports a model with block structure (assigned via ``Model.blocks``) into the doubly-bordered block-diagonal ("arrowhead") format consumed by the distributed solver `PIPS-IPM++ <https://gitlab.com/pips-ipmpp/pips-ipmpp>`_, deriving the submatrices from linopy's internal CSR. ``linopy.io.read_pips_files`` reconstructs an equivalent flat model (usable to validate the export against any solver with no PIPS build), and ``read_pips_solution`` reads the solver output back. See :doc:`pips-installation`.
 * An env-gated ``pips`` solver (``solver_name="pips"``, enabled through the ``PIPS_BINARY`` environment variable) runs the export through PIPS under MPI and maps the primal, duals and objective back onto the model. It stays absent from ``available_solvers`` unless the driver binary is present, so ordinary installs are unaffected.
-* New ``linopy.pips`` HPC helpers: ``assign_blocks(m, dim, n_blocks)`` (also ``Model.assign_blocks``) splits a dimension into contiguous blocks, and ``diagnose(m)`` reports the arrowhead structure — border fraction, block balance, adjacent-vs-border linking rows and a rank/thread recommendation — to judge whether a decomposition will scale before running. ``PipsConfig`` controls the parallel launch (``launcher``, ``n_ranks``, ``threads_per_rank``, ``linear_solver``) via ``solver_options`` or environment variables, and ``write_job`` emits a SLURM batch script for detached cluster runs.
+* New ``linopy.backends.pips`` HPC helpers: ``assign_blocks(m, dim, n_blocks)`` (also ``Model.assign_blocks``) splits a dimension into contiguous blocks, and ``diagnose(m)`` reports the arrowhead structure — border fraction, block balance, adjacent-vs-border linking rows and a rank/thread recommendation — to judge whether a decomposition will scale before running. ``PipsConfig`` controls the parallel launch (``launcher``, ``n_ranks``, ``threads_per_rank``, ``linear_solver``) via ``solver_options`` or environment variables, and ``write_job`` emits a SLURM batch script for detached cluster runs.
 
 *Other*
