+++ linopy/backends/__init__.py
@@ -1,0 +1,1 @@
+"""Backend-specific integrations that need more than a solver class in `solvers.py`."""
